requirements.txt
PySide6==6.8.0.2
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
pywinrm==0.5.0
